# frontend/src/features/research-universe/openapi.json
{
  "components": {
    "schemas": {
...
      "AnchorCommand": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "end": {
            "exclusiveMinimum": 0.0,
            "title": "End",
            "type": "integer"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "note_id": {
            "title": "Note Id",
            "type": "string"
          },
          "note_revision_id": {
            "title": "Note Revision Id",
            "type": "string"
          },
          "selected_text": {
            "minLength": 1,
            "title": "Selected Text",
            "type": "string"
          },
          "start": {
            "minimum": 0.0,
            "title": "Start",
            "type": "integer"
          }
        },
        "required": [
          "command_id",
          "expected_sequence",
          "note_id",
          "note_revision_id",
          "start",
          "end",
          "selected_text"
        ],
        "title": "AnchorCommand",
        "type": "object"
      },
...
      "ClaimCommand": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "kind": {
            "anyOf": [
              {
                "enum": [
                  "consensus",
                  "division",
                  "vacancy"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "expected_sequence",
          "text"
        ],
        "title": "ClaimCommand",
        "type": "object"
      },
      "Command": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          }
        },
        "required": [
          "command_id",
          "expected_sequence"
        ],
        "title": "Command",
        "type": "object"
      },
      "CommandResponse": {
        "properties": {
          "commit_position": {
            "title": "Commit Position",
            "type": "integer"
          },
          "event_ids": {
            "items": {
              "type": "string"
            },
            "title": "Event Ids",
            "type": "array"
          },
          "fragment": {
            "additionalProperties": true,
            "title": "Fragment",
            "type": "object"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          }
        },
        "required": [
          "commit_position",
          "event_ids",
          "result",
          "fragment"
        ],
        "title": "CommandResponse",
        "type": "object"
      },
...
      "CorpusSearchHit": {
        "properties": {
          "matched_terms": {
            "title": "Matched Terms",
            "type": "integer"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "source_locator": {
            "title": "Source Locator",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "material_id",
          "source_locator",
          "title",
          "matched_terms",
          "snippet"
        ],
        "title": "CorpusSearchHit",
        "type": "object"
      },
      "CorpusSearchResponse": {
        "properties": {
          "group": {
            "title": "Group",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CorpusSearchHit"
            },
            "title": "Results",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "query",
          "group",
          "total",
          "results"
        ],
        "title": "CorpusSearchResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NoteCommand": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "expected_sequence",
          "text"
        ],
        "title": "NoteCommand",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceCommand": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "expected_sequence": {
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "expected_sequence",
          "question"
        ],
        "title": "WorkspaceCommand",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cui native test",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/claims/{claim_id}/review-rounds": {
      "post": {
        "operationId": "round_api_v2_claims__claim_id__review_rounds_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Round",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
    "/api/v2/corpus/search": {
      "get": {
        "operationId": "search_api_v2_corpus_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "legacy"
              ],
              "title": "Group",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "research-universe-corpus-search"
        ]
      }
    },
...
    "/api/v2/review-rounds/{round_id}": {
      "get": {
        "operationId": "review_round_api_v2_review_rounds__round_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Round",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
...
    "/api/v2/universes/{universe_id}/home": {
      "get": {
        "operationId": "home_api_v2_universes__universe_id__home_get",
        "parameters": [
          {
            "in": "path",
            "name": "universe_id",
            "required": true,
            "schema": {
              "title": "Universe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Home",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
    "/api/v2/universes/{universe_id}/workspaces": {
      "post": {
        "operationId": "create_workspace_api_v2_universes__universe_id__workspaces_post",
        "parameters": [
          {
            "in": "path",
            "name": "universe_id",
            "required": true,
            "schema": {
              "title": "Universe Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workspace",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
...
    "/api/v2/workspaces/{workspace_id}": {
      "get": {
        "operationId": "workspace_api_v2_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
...
    "/api/v2/workspaces/{workspace_id}/anchors": {
      "post": {
        "operationId": "anchor_api_v2_workspaces__workspace_id__anchors_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnchorCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anchor",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
...
    "/api/v2/workspaces/{workspace_id}/claims": {
      "post": {
        "operationId": "claim_api_v2_workspaces__workspace_id__claims_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
...
    "/api/v2/workspaces/{workspace_id}/notes": {
      "post": {
        "operationId": "save_note_api_v2_workspaces__workspace_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Note",
        "tags": [
          "research-universe-slice1"
        ]
      }
    },
...
  }
}